wechat-ocr-api/requirements.txt
google-cloud-vision==3.4.1
flask==2.3.2